+++ test/extended_math_test.dart
@@ -4,13 +4,14 @@ import 'package:extended_math/extended_math.dart';
 
 void main() {
   final i1 = Matrix(<List<num>>[
-    <int>[1, 0, 0, 0],
-    <int>[0, 0, 0, 0],
-    <int>[1, -7, 0, 0]
+    <int>[0, 0, 1, 0],
+    <int>[0, 0, 0, 2],
+    <int>[0, 0, 0, 0]
   ]);
 
-  final i2 = SquareMatrix(<List<num>>[
+  final i2 = Matrix(<List<num>>[
     <int>[3, 2, 2],
+    <int>[0, 3, 1],
     <int>[0, 3, 1],
     <int>[0, 3, 6]
   ]);
@@ -18,6 +19,5 @@ void main() {
   final v1 = Vector(<int>[6, 5, 7]);
   final v2 = Vector(<int>[6, 3, 1]);
 
-  print(i1.rowAt(1).removeLast());
-  print(i1);
+  print(i1.matrixProduct(i2));
 }